# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Onboarding API
  description: API for managing customer account applications for ALEXA Bank.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Development Server

paths:
  /applications:
    post:
      tags:
        - Applications
      summary: Start a new application
      description: Creates a new account application in DRAFT status. Returns the application ID to be used for subsequent updates.
      operationId: createApplication
      requestBody:
        description: Initial application data. Name, Address, and DOB are required.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationCreateRequest'
      responses:
        '201':
          description: Application created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                status: "DRAFT"
  
  /applications/{id}:
    get:
      tags:
        - Applications
      summary: Get application details
      description: Retrieves the current state of an application. Use this to 'Resume' a paused application.
      operationId: getApplication
      parameters:
        - name: id
          in: path
          required: true
          description: The Application ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Application found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '404':
          description: Application not found

    patch:
      tags:
        - Applications
      summary: Update application details
      description: Updates specific fields of the application. Per-field validation is performed immediately. If any field is invalid, a 400 error with details is returned. Valid fields are saved (Implicit Pause).
      operationId: updateApplication
      parameters:
        - name: id
          in: path
          required: true
          description: The Application ID
          schema:
            type: string
            format: uuid
      requestBody:
        description: Partial application data to update.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationUpdateRequest'
      responses:
        '200':
          description: Application updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '400':
          description: Validation failed for one or more fields
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Application not found
        '409':
          description: Application is already completed and cannot be modified

  /applications/{id}/submit:
    post:
      tags:
        - Applications
      summary: Submit application
      description: Finalizes the application. Performs a full completeness check. If all mandatory fields are present, status changes to COMPLETED and a confirmation is returned.
      operationId: submitApplication
      parameters:
        - name: id
          in: path
          required: true
          description: The Application ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Application submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '400':
          description: Application is incomplete (missing mandatory fields)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Application not found

components:
  schemas:
    AccountType:
      type: string
      enum: [SAVINGS, CURRENT, INVESTMENT]
      description: The type of bank account being requested. Each type has different features and interest rates.
      example: "SAVINGS"
    
    ApplicationStatus:
      type: string
      enum: [DRAFT, COMPLETED]
      description: |
        The current status of the application lifecycle:
        * `DRAFT` - The application is in progress and can be updated.
        * `COMPLETED` - The application has been submitted and finalized. It cannot be modified.
      example: "DRAFT"

    Address:
      type: object
      description: Reusable Address component.
      properties:
        streetName:
          type: string
          description: Street name.
          example: "Kalverstraat"
        houseNumber:
          type: string
          description: House number including any additions.
          example: "101"
        postCode:
          type: string
          description: Dutch Postcode. Must consist of 4 digits followed by an optional space and 2 letters.
          pattern: '^\d{4}\s?[a-zA-Z]{2}$'
          x-field-extra-annotation: "@jakarta.validation.constraints.Pattern(regexp = \"^\\\\d{4}\\\\s?[a-zA-Z]{2}$\")"
          example: "1012 PX"
        city:
          type: string
          description: City or town name.
          example: "Amsterdam"

    ApplicationCreateRequest:
      type: object
      description: |
        Data required to start a new application.
        Name, Address, and Date of Birth are mandatory to create a draft.
      required:
        - name
        - address
        - dateOfBirth
      properties:
        name:
          type: string
          description: Full legal name of the applicant.
          minLength: 2
          maxLength: 100
          example: "Johan Doe"
        address:
          $ref: '#/components/schemas/Address'
        dateOfBirth:
          type: string
          format: date
          description: Applicant's date of birth (YYYY-MM-DD). Must be in the past.
          x-field-extra-annotation: "@jakarta.validation.constraints.Past"
          example: "1990-05-20"
        email:
          type: string
          format: email
          description: Valid email address for contact and notifications.
          x-field-extra-annotation: "@jakarta.validation.constraints.Email"
          example: "johan.doe@example.com"
        idDocument:
          type: string
          description: Document number of the applicant's identification.
          example: "NL123456789"
        accountType:
          $ref: '#/components/schemas/AccountType'

    ApplicationUpdateRequest:
      type: object
      description: |
        Data transfer object for updating an account application.
        All fields are optional to support partial updates (PATCH).
      properties:
        name:
          type: string
          description: Full legal name of the applicant.
          minLength: 2
          maxLength: 100
          example: "Johan Doe"
        address:
          $ref: '#/components/schemas/Address'
        dateOfBirth:
          type: string
          format: date
          description: Applicant's dateOfBirth (YYYY-MM-DD). Must be in the past.
          x-field-extra-annotation: "@jakarta.validation.constraints.Past"
          example: "1990-05-20"
        idDocument:
          type: string
          description: Document number of the applicant's identification.
          example: "NL123456789"
        accountType:
          $ref: '#/components/schemas/AccountType'
        startingBalance:
          type: number
          minimum: 0
          description: The initial amount to be deposited into the account upon opening.
          x-field-extra-annotation: "@jakarta.validation.constraints.DecimalMin(value = \"0.0\", inclusive = true)"
          example: 1000.00
        monthlySalary:
          type: number
          minimum: 0
          description: The applicant's gross monthly salary.
          x-field-extra-annotation: "@jakarta.validation.constraints.DecimalMin(value = \"0.0\", inclusive = true)"
          example: 3500.00
        interestedInOtherProducts:
          type: boolean
          description: Indicates if the applicant is interested in receiving information about insurance or credit products.
          example: true
        email:
          type: string
          format: email
          description: Valid email address for contact and notifications.
          x-field-extra-annotation: "@jakarta.validation.constraints.Email"
          example: "johan.doe@example.com"
    
    ApplicationResponse:
      type: object
      description: Full details of an Account Application, including system-generated fields.
      properties:
        id:
          type: string
          format: uuid
          description: Unique UUID assigned to the application upon creation.
          example: "550e8400-e29b-41d4-a716-446655440000"
        status:
          $ref: '#/components/schemas/ApplicationStatus'
        name:
          type: string
          description: Full legal name of the applicant.
          example: "Johan Doe"
        address:
          $ref: '#/components/schemas/Address'
        dateOfBirth:
          type: string
          format: date
          description: Date of Birth.
          example: "1990-05-20"
        idDocument:
          type: string
          description: ID Document Number.
          example: "NL123456789"
        accountType:
          $ref: '#/components/schemas/AccountType'
        startingBalance:
          type: number
          description: Initial deposit amount.
          example: 1000.00
        monthlySalary:
          type: number
          description: Monthly salary income.
          example: 3500.00
        interestedInOtherProducts:
          type: boolean
          description: Interest in other products.
          example: true
        email:
          type: string
          format: email
          description: Contact email address.
          example: "johan.doe@example.com"
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the application was created.
          example: "2023-10-27T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the application was last updated.
          example: "2023-10-27T10:15:00Z"

    ProblemDetail:
      type: object
      description: RFC 7807 Standard Error Response.
      properties:
        type:
          type: string
          description: URI reference that identifies the problem type.
          example: "about:blank"
        title:
          type: string
          description: Short, human-readable summary of the problem type.
          example: "Bad Request"
        status:
          type: integer
          description: The HTTP status code generated by the origin server.
          example: 400
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence of the problem.
          example: "Validation failed for object='accountApplication'. Error count: 1"
        instance:
          type: string
          description: URI reference that identifies the specific occurrence of the problem.
          example: "/applications/550e8400-e29b-41d4-a716-446655440000/submit"
        properties:
          type: object
          description: Additional properties providing specific error details (e.g., field validation errors).
          additionalProperties: true